FDG-11431 set api error on metadata fetch
