+++ src/components/dtg-table/dtg-table.jsx
@@ -147,8 +147,7 @@ export default function DtgTable({
       loadTimer = setTimeout(() => loadingTimeout(loadCanceled, setIsLoading), netLoadingDelay);
       const { from, to } = getDateFilters(filteredDateRange, dateRange);
       const startPage = resetPage ? 1 : currentPage;
-      let sortData;
-      sortData = tableMeta?.meta && table && sorting?.length > 0 ? getSortedColumnsData(table, hideColumns, tableMeta.meta.dataTypes) : [];
+      const sortData = tableMeta?.meta && table && sorting?.length > 0 ? getSortedColumnsData(table, hideColumns, tableMeta.meta.dataTypes) : [];
       setTableColumnSortData(sortData);
       pagedDatatableRequest(
         selectedTable,
@@ -331,67 +330,63 @@ export default function DtgTable({
   return (
     <div className={overlayContainer}>
       {/* Loading Indicator */}
-      {!isLoading && !reactTableData && !selectedTable?.apiFilter && <LoadingIndicator loadingClass={loadingIcon} overlayClass={overlay} />}
+      {!isLoading && !reactTableData && !selectedTable?.apiFilter && !(apiError || tableProps.apiError) && (
+        <LoadingIndicator loadingClass={loadingIcon} overlayClass={overlay} />
+      )}
       {/* Data Dictionary and Dataset Detail tables */}
       <div data-testid="table-content">
         {/* API Error Message */}
-        {(apiError || tableProps.apiError) && !emptyDataMessage && (
-          <>
-            <DtgTableApiError />
-          </>
-        )}
+        {(apiError || tableProps.apiError) && !emptyDataMessage && <DtgTableApiError />}
         {reactTableData?.data && (
-          <ErrorBoundary FallbackComponent={() => <></>}>
-            <>
-              <div data-testid="table-content" className={overlayContainerNoFooter}>
-                <div className={selectColumnsWrapper}>
-                  {defaultSelectedColumns && (
-                    <div className={selectColumnPanel ? selectColumnPanelActive : selectColumnPanelInactive} data-testid="selectColumnsMainContainer">
-                      <DataTableColumnSelector
-                        selectColumnPanel={selectColumnPanel}
-                        fields={allColumns}
-                        resetToDefault={() => setColumnVisibility(getInvisibleColumns(defaultSelectedColumns, allColumns))}
-                        setSelectColumnPanel={setSelectColumnPanel}
-                        defaultSelectedColumns={defaultSelectedColumns}
+          <ErrorBoundary FallbackComponent={() => <DtgTableApiError />}>
+            <div data-testid="table-content" className={overlayContainerNoFooter}>
+              <div className={selectColumnsWrapper}>
+                {defaultSelectedColumns && (
+                  <div className={selectColumnPanel ? selectColumnPanelActive : selectColumnPanelInactive} data-testid="selectColumnsMainContainer">
+                    <DataTableColumnSelector
+                      selectColumnPanel={selectColumnPanel}
+                      fields={allColumns}
+                      resetToDefault={() => setColumnVisibility(getInvisibleColumns(defaultSelectedColumns, allColumns))}
+                      setSelectColumnPanel={setSelectColumnPanel}
+                      defaultSelectedColumns={defaultSelectedColumns}
+                      table={table}
+                      additionalColumns={additionalColumns}
+                      defaultColumns={defaultColumns}
+                    />
+                  </div>
+                )}
+                <div className={tableStyle}>
+                  <div data-test-id="table-content" className={rawDataTableContainer}>
+                    <table {...tableProps.aria}>
+                      <DataTableHeader
                         table={table}
-                        additionalColumns={additionalColumns}
-                        defaultColumns={defaultColumns}
+                        dataTypes={reactTableData.meta.dataTypes}
+                        resetFilters={resetFilters}
+                        manualPagination={manualPagination}
+                        allActiveFilters={allActiveFilters}
+                        setAllActiveFilters={setAllActiveFilters}
+                        disableDateRangeFilter={disableDateRangeFilter}
                       />
-                    </div>
-                  )}
-                  <div className={tableStyle}>
-                    <div data-test-id="table-content" className={rawDataTableContainer}>
-                      <table {...tableProps.aria}>
-                        <DataTableHeader
-                          table={table}
-                          dataTypes={reactTableData.meta.dataTypes}
-                          resetFilters={resetFilters}
-                          manualPagination={manualPagination}
-                          allActiveFilters={allActiveFilters}
-                          setAllActiveFilters={setAllActiveFilters}
-                          disableDateRangeFilter={disableDateRangeFilter}
-                        />
-                        <DataTableBody
-                          table={table}
-                          dataTypes={reactTableData.meta.dataTypes}
-                          detailViewConfig={config?.detailView}
-                          setDetailViewState={setDetailViewState}
-                          setSummaryValues={setSummaryValues}
-                        />
-                      </table>
-                    </div>
+                      <DataTableBody
+                        table={table}
+                        dataTypes={reactTableData.meta.dataTypes}
+                        detailViewConfig={config?.detailView}
+                        setDetailViewState={setDetailViewState}
+                        setSummaryValues={setSummaryValues}
+                      />
+                    </table>
                   </div>
                 </div>
               </div>
-              <DataTableFooter
-                table={table}
-                showPaginationControls={showPaginationControls}
-                pagingProps={pagingProps}
-                manualPagination={manualPagination}
-                rowsShowing={rowsShowing}
-                setTableDownload={setTableRowSizeData}
-              />
-            </>
+            </div>
+            <DataTableFooter
+              table={table}
+              showPaginationControls={showPaginationControls}
+              pagingProps={pagingProps}
+              manualPagination={manualPagination}
+              rowsShowing={rowsShowing}
+              setTableDownload={setTableRowSizeData}
+            />
           </ErrorBoundary>
         )}
       </div>
